smoke: rework shaders & optimize regions
